src/backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Nikita Nelson <nikitazharov2@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
pytest = "^7.2.0"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
SQLAlchemy = "^1.4.44"
alembic = "^1.8.1"
psycopg2 = "^2.9.5"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start="app.main:start"
